feat: enables the submission related functions of the api
This commit enables users to upload submissions, check submissions'
list, results, and details. It also allows admins to rejudge
submissions. The submit code part is tested partailly. The system
responded correctly when a user submitted a correct solution.
However, the other functions still needed to be tested, and there
is also the need for pressure test