test: tested all known judge conditions
Tested all known judge conditions with cpp since python has not compile
error. Finding out that the error of devide by 0 is judged as an UNK when
is should be an RE. It would be fixed in the next commit. In addition,
due to the speed of cpp, it's hard to make a test case that really
exceeds the time limit. Therefore, the test came up to be AC at last.