fix: fixed the issue of api rebooting during judging the code
This commit fixed the issuf of api often rebooting during the code
judging process. This happened because the original tmp directory,
which was the code judging file, was inside the app folder. When
there were changes in that directory, fastpi would reboot itself.
In addition, it was not very reasonable to put this directory in
the project folder. Therefore, it is now moved into the /tmp directory.