fix: fixed the issue of cannot delete all related cache by id totally
This commit fixed the issue of when deleting cache by the id key map,
the cache with status code that is not 200 would still exists since
they were not related to any submission. However, this should be only
a temporary solution since a better way to solve this is to raise
httpexception in the function. However, I did not use that solution
since all endpoints codes used the way of using the response as a
parameter, and its status would be changed if there was something
wrong. This sould be improved when updating the returning way of
all endpoints