refactor: rewrote the get_current_user function
This commit refactor the get_current_user function. In the past, it can
only raise 401 when the user is not logged in. Now, it is built by the
factory function get_current_user_factory(admin_only: bool). According to
the admin_only param, it will return a get_current_user function that will
raise 403 if the current is not an admin correspondingly.

This function has passed the previous two tests.