test: tested all problems endpoints
This commit tested all problems endpoints. During the tests, a issue of the
problem list endpoint is discovered. The issue is that the endpoint will return
bad request when the hardness param is None. It is fixed by the typename optional
and a default value for the query.

All tests succeeded after fixing.